+++ clang/lib/Basic/Targets/SystemZ.h
@@ -105,6 +105,7 @@ public:
     HasFastHalfType = false;
 
     HasStrictFP = true;
+    HasSignalingNaNs = true;
   }
 
   unsigned getMinGlobalAlign(uint64_t Size, bool HasNonWeakDef) const override;