+++ llvm/lib/Transforms/Utils/CodeExtractor.cpp
@@ -1006,6 +1006,7 @@ Function *CodeExtractor::constructFunctionDeclaration(
       case Attribute::SanitizeRealtime:
       case Attribute::SanitizeRealtimeBlocking:
       case Attribute::SanitizeAllocToken:
+      case Attribute::SignalingNans:
       case Attribute::SpeculativeLoadHardening:
       case Attribute::StackProtect:
       case Attribute::StackProtectReq: