+++ clang/docs/ReleaseNotes.md
@@ -115,6 +115,9 @@ latest release, please see the [Clang Web Site](https://clang.llvm.org) or the
 
 ### New Compiler Flags
 
+- New option `-f[no-]signaling-nans` added to control the support of
+  signaling NaNs.
+
 ### Deprecated Compiler Flags
 
 ### Modified Compiler Flags